client: fix compiler warnings on Mac and Ubuntu
svn path=/trunk/boinc/; revision=24034

+++ client/coproc_detect.cpp
@@ -308,7 +308,8 @@ void COPROCS::get_opencl(bool use_all, vector<string>&warnings,
                     i, nvidia_opencls[i].name, i, nvidia_opencls[i].device_id, i, nvidia_opencls[i].device_num);
                 msg_printf(0, MSG_INFO,
                     "[coproc_debug] COPROC_NVIDIA [no CUDA]: nvidia_opencls[%d].global_RAM = %lu; nvidia_opencls[%d].local_RAM = %lu",
-                    i, nvidia_opencls[i].global_RAM, i, nvidia_opencls[i].local_RAM);
+                    i, (unsigned long)nvidia_opencls[i].global_RAM, 
+                    i, (unsigned long)nvidia_opencls[i].local_RAM);
             }
             if (in_vector(nvidia_opencls[i].device_num, ignore_nvidia_dev)) continue;
             bool is_best = false;
@@ -384,7 +385,8 @@ void COPROCS::get_opencl(bool use_all, vector<string>&warnings,
                     i, ati_opencls[i].name, i, ati_opencls[i].device_id, i, ati_opencls[i].device_num);
                 msg_printf(0, MSG_INFO,
                     "[coproc_debug] COPROC_ATI [no CAL]: ati_opencls[%d].global_RAM = %lu; ati_opencls[%d].local_RAM = %lu",
-                    i, ati_opencls[i].global_RAM, i, ati_opencls[i].local_RAM);
+                    i, (unsigned long)ati_opencls[i].global_RAM, 
+                    i, (unsigned long)ati_opencls[i].local_RAM);
             }
             if (in_vector(ati_opencls[i].device_num, ignore_ati_dev)) continue;
             bool is_best = false;
@@ -1088,7 +1090,7 @@ bool COPROC_NVIDIA::matches(OPENCL_DEVICE_PROP& OpenCLprop) {
             device_num, prop.deviceHandle, OpenCLprop.device_id);
         msg_printf(0, MSG_INFO,
             "[coproc_debug] COPROC_NVIDIA [in matches()]: prop.totalGlobalMem = %u; OpenCLprop.global_RAM = %lu; OpenCLprop.local_RAM = %lu",
-            prop.totalGlobalMem, OpenCLprop.global_RAM, OpenCLprop.local_RAM);
+            prop.totalGlobalMem, (unsigned long)OpenCLprop.global_RAM, (unsigned long)OpenCLprop.local_RAM);
 
         if (strcmp(prop.name, OpenCLprop.name)) {
             msg_printf(0, MSG_INFO,
@@ -1551,7 +1553,7 @@ bool COPROC_ATI::matches(OPENCL_DEVICE_PROP& OpenCLprop) {
             name, OpenCLprop.name);
         msg_printf(0, MSG_INFO,
             "[coproc_debug] COPROC_ATI [in matches()]: attribs.localRAM = %u; OpenCLprop.global_RAM = %lu; OpenCLprop.local_RAM = %lu",
-            attribs.localRAM, OpenCLprop.global_RAM, OpenCLprop.local_RAM);
+            attribs.localRAM, (unsigned long)OpenCLprop.global_RAM, (unsigned long)OpenCLprop.local_RAM);
     }
 
 #if 0//def _WIN32
